\encoding{UTF-8}
\name{milk}
\alias{milk}
\docType{data}
\title{
Average daily fat yields.

}
\description{
This data is the average daily fat yields (kg/day) of milk from a single cow for each of 35 weeks; see Kokonendji et al. (2009).

}
\usage{data(milk)}
\format{
  A data frame with 35 observations on the following 2 variables.
  \describe{
    \item{\code{week}}{Number of the week}
    \item{\code{yield}}{The yield quantity}
  }
}

\source{ 
 McCulloch, C.E. (2001). An Introduction to Generalized Linear Mixed Models,
46a Reuniao Anual da RBRAS - 9o SEAGRO, University of Sao Paulo -
ESALQ, Piracicaba. 

}
\references{

Kokonendji, C.C., Senga Kiessé, T. and Demétrio, C.G.B. (2009). Appropriate kernel regression on a count explanatory variable and applications,
\emph{Advances and Applications in Statistics}  \bold{12},  99 - 125.

}
\examples{
data(milk)

}
\keyword{datasets}
\keyword{datasets}
