\name{plot.hcvc.fun}
\alias{plot.hcvc.fun}
\alias{plot.hcvd.fun}

\title{Plot of cross-validation function for bandwidth selection in density or p.m.f. estimation.

}
\description{ The functions allows to plot the cross-validation both in discrete \code{plot.hcvd.fun} and continuous \code{plot.hcvc.fun} cases.

}
\usage{
\method{plot}{hcvc.fun}(x,  \dots)
\method{plot}{hcvd.fun}(x,  \dots)

}

\arguments{
  \item{x}{ an object

}
  \item{\dots}{ Further arguments

}
}
\details{  Plot a graphic for cross-validation function

}
\value{ returns a graphics

}
\references{
Kokonendji, C.C. and Senga Kiessé, T. (2011). Discrete associated kernel method and extensions,
\emph{Statistical Methodology}  \bold{8},  497 - 516.

Kokonendji, C.C., Senga Kiessé, T. and Zocchi, S.S. (2007). Discrete triangular distributions and non-parametric estimation for
probability mass function.
\emph{Journal of Nonparametric Statistics}  \bold{19},  241 - 254.

Zougab, N., Adjabi, S. and Kokonendji, C.C. (2012). Binomial kernel and Bayes local bandwidth in discrete functions estimation. 
\emph{Journal of Nonparametric Statistics}  \bold{24}, 783 - 795.

}
\author{
W. E. Wansouwé, S. M. Somé and C. C. Kokonendji 
}

