\name{plot.kern.fun}
\alias{plot.kern.fun}


\title{ Plot of  associated kernel function

}
\description{ The \code{\link{plot.kern.fun}} function loops through calls to 
the \code{\link{kern.fun}} function. 

}
\usage{
\method{plot}{kern.fun}(x, \dots)

}

\arguments{
  \item{x}{an object of class \code{kern.fun} (output from \code{\link{kern.fun}}).

}
  \item{\dots}{
Other graphics parameters

}
}
\value{
  Plot of associated the  kernel function is sent to graphics window.
}
\references{

Kokonendji, C.C. and Senga Kiessé, T. (2011). Discrete associated kernel method and extensions,
 \emph{Statistical Methodology}  \bold{8},  497 - 516.

Kokonendji, C.C., Senga Kiessé, T. and Demétrio, C.G.B. (2009). Appropriate kernel regression on a count explanatory variable and applications,
\emph{Advances and Applications in Statistics}  \bold{12},  99 - 125.

}
\author{
W. E. Wansouwé, S. M. Somé and C. C. Kokonendji 
}

