\name{plot.kpmfe.fun}
\alias{plot.kpmfe.fun}


\title{
Plot of the function for associated kernel estimation of the p.m.f.
}
\description{ The function plots the p.m.f. estimation in a single value or in a
grid using discrete associated kernels. Three different associated kernels are available: DiracDU (for categorical data), binomial and discrete triangular (for count data).

}
\usage{
\method{plot}{kpmfe.fun}(x, \dots)
}

\arguments{
  \item{x}{ An object of class \code{kpmfe.fun}.

}
  \item{\dots}{ Further arguments

}
}
\details{ Plot a graphic

}
\references{
Kokonendji, C.C. and Senga Kiessé, T. (2011). Discrete associated kernel method and extensions,
\emph{Statistical Methodology}  \bold{8},  497 - 516.

Kokonendji, C.C., Senga Kiessé, T. and Zocchi, S.S. (2007). Discrete triangular distributions and non-parametric estimation for
probability mass function.
\emph{Journal of Nonparametric Statistics}  \bold{19},  241 - 254.

Zougab, N., Adjabi, S. and Kokonendji, C.C. (2012). Binomial kernel and Bayes local bandwidth in discrete functions estimation. 
\emph{Journal of Nonparametric Statistics}  \bold{24}, 783 - 795.

}
\author{
W. E. Wansouwé, S. M. Somé and C. C. Kokonendji 
}

