\name{plot.reg.fun}
\alias{plot.reg.fun}


\title{
Plot for associated kernel regression
}
\description{ Plot for associated kernel regression for univariate data. The \code{\link{plot.reg.fun}} function loops through calls to 
the \code{\link{reg.fun}} function. 

}
\usage{
\method{plot}{reg.fun}(x,  \dots)

}

\arguments{
  \item{x}{ An object of class \code{reg.fun}

}
  \item{\dots}{ other graphics parameters 
}
}
\details{ The function allows to plot the regression

}
\value{ Plot is sent to graphics window.

}
\references{

Kokonendji, C.C. and Senga Kiessé, T. (2011). Discrete associated kernel method and extensions,
\emph{Statistical Methodology}  \bold{8},  497 - 516.

Kokonendji, C.C., Senga Kiessé, T. and Zocchi, S.S. (2007). Discrete triangular distributions and non-parametric estimation for
probability mass function.
\emph{Journal of Nonparametric Statistics}  \bold{19},  241 - 254.

Zougab, N., Adjabi, S. and Kokonendji, C.C. (2012). Binomial kernel and Bayes local bandwidth in discrete functions estimation. 
\emph{Journal of Nonparametric Statistics}  \bold{24}, 783 - 795.

}
\author{
W. E. Wansouwé, S. M. Somé and C. C. Kokonendji  
}

